# openapi.yaml
openapi: 3.1.0
info:
  title: SEMD API
  version: 1.0.0
paths:
...
  /report:
    get:
      tags:
      - report
      summary: Get All Reports
      description: Get all URL reports. Requires authentication.
      operationId: get_report_data_report_get
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportListResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
    post:
      tags:
      - report
      summary: Create Report
      description: Create a new URL report. Requires authentication.
      operationId: create_report_report_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UrlReportCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /report/me:
    get:
      tags:
      - report
      summary: Get My Reports
      description: Get URL reports created by the current user.
      operationId: get_my_reports_report_me_get
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportListResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /report/{report_id}:
    get:
      tags:
      - report
      summary: Get Report by ID
      description: Get a specific URL report by its ID.
      operationId: get_report_by_id_report__report_id__get
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: integer
          title: Report Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
    put:
      tags:
      - report
      summary: Update Report
      description: Update an existing URL report. Requires authentication.
      operationId: update_report_report__report_id__put
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: integer
          title: Report Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UrlReportUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /report/{report_id}/history:
    get:
      tags:
      - report
      summary: Get Report History
      description: Get the action history of a specific URL report.
      operationId: get_report_history_report__report_id__history_get
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: integer
          title: Report Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportListResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
...
  /stat/report:
    get:
      tags:
      - stat-report
      summary: Get Report Stat
      operationId: get_report_stat_stat_report_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportStatResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
...
components:
  schemas:
...
    FlagType:
      type: string
      enum:
      - MALICIOUS
      - BENIGN
      title: FlagType
...
    ReportListResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          items: {}
          type: array
          title: data
          description: List of reports
      type: object
      required:
      - status
      - message
      - data
      title: ReportListResponse
    ReportResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          title: data
          description: Report data
      type: object
      required:
      - status
      - message
      - data
      title: ReportResponse
    ReportStatItem:
      properties:
        reported_by_amount:
          type: integer
          title: reported_by_amount
          description: "\u0E25\u0E34\u0E49\u0E07\u0E04\u0E4C\u0E17\u0E35\u0E16\u0E39\
            \u0E01\u0E23\u0E32\u0E22\u0E07\u0E32\u0E19\u0E42\u0E14\u0E22\u0E04\u0E38\
            \u0E13"
        safe_urls:
          type: integer
          title: safe_urls
          description: "\u0E25\u0E34\u0E49\u0E07\u0E04\u0E4C\u0E1B\u0E25\u0E2D\u0E14\
            \u0E20\u0E31\u0E22"
        malicious_urls:
          type: integer
          title: malicious_urls
          description: "\u0E25\u0E34\u0E49\u0E07\u0E04\u0E4C\u0E2D\u0E31\u0E19\u0E15\
            \u0E23\u0E32\u0E22"
        error_urls:
          type: integer
          title: error_urls
          description: "\u0E23\u0E32\u0E22\u0E07\u0E32\u0E19\u0E1C\u0E34\u0E14\u0E1E\
            \u0E25\u0E32\u0E14"
        pending_urls:
          type: integer
          title: pending_urls
          description: "\u0E25\u0E34\u0E49\u0E07\u0E04\u0E4C\u0E23\u0E2D\u0E15\u0E23\
            \u0E27\u0E08\u0E2A\u0E2D\u0E1A"
      type: object
      required:
      - reported_by_amount
      - safe_urls
      - malicious_urls
      - error_urls
      - pending_urls
      title: ReportStatItem
...
    ReportStatResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          $ref: '#/components/schemas/ReportStatItem'
      type: object
      required:
      - status
      - message
      - data
      title: ReportStatResponse
...
    ReportStatusType:
      type: string
      enum:
      - PENDING
      - ACCEPTED
      - REJECTED
      title: ReportStatusType
...
    UrlReportCreateRequest:
      properties:
        url:
          type: string
          title: Url
          description: URL to report
        categories:
          $ref: '#/components/schemas/FlagType'
          description: Category of the URL (MALICIOUS or BENIGN)
          default: BENIGN
        status:
          $ref: '#/components/schemas/ReportStatusType'
          description: Status of the report
          default: PENDING
        remark:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Remark
          description: Additional remark
      type: object
      required:
      - url
      title: UrlReportCreateRequest
    UrlReportUpdateRequest:
      properties:
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
          description: URL to report
        categories:
          anyOf:
          - $ref: '#/components/schemas/FlagType'
          - type: 'null'
          description: Category of the URL (MALICIOUS or BENIGN)
        status:
          anyOf:
          - $ref: '#/components/schemas/ReportStatusType'
          - type: 'null'
          description: Status of the report
        remark:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Remark
          description: Additional remark
      type: object
      title: UrlReportUpdateRequest
...